Update pupl_interp.m
{ s = find([interpidx(1),diff(interpidx) == 1]); e = find([diff(interpidx) == -1, interpidx(end)]); }
This to get all the start and end index with their length always the same, so i think we don't need the
{ if s(1) > e(1), e(1) = []; end,  if s(end) > e(end), s(end) = []; end } anymore.

{ if s(1) == 1, interpidx(s(1):e(1)) = false; , elseif e(end) == npoints, interpidx(s(end):e(end)) = false; end}
This is to make cubic interpolation skip the begining and last nan cluster to avoid make negative value. 

By the way, i think pupl is bettter than chap and gazer, thank you for your work.